+++ twitter-langchain/twitter_langchain/__init__.py
@@ -1,11 +1,11 @@
 """CDP Twitter Toolkit."""
 
-from twitter_langchain.twitter_action import TwitterAction
 from twitter_langchain.twitter_api_wrapper import TwitterApiWrapper
+from twitter_langchain.twitter_tool import TwitterTool
 from twitter_langchain.twitter_toolkit import TwitterToolkit
 
 __all__ = [
-    "TwitterAction",
     "TwitterApiWrapper",
+    "TwitterTool",
     "TwitterToolkit",
 ]