+++ twitter-langchain/twitter_langchain/twitter_tool.py
@@ -1,5 +1,15 @@
-"""Tool allows agents to interact with the Twitter API."""
+"""Tool allows agents to interact with the Twitter API.
 
+To use this tool, you must first set as environment variables:
+    OPENAI_API_KEY
+    TWITTER_API_KEY
+    TWITTER_API_SECRET
+    TWITTER_ACCESS_TOKEN
+    TWITTER_ACCESS_TOKEN_SECRET
+
+"""
+
+from collections.abc import Callable
 from typing import Any
 
 from langchain_core.callbacks import CallbackManagerForToolRun
@@ -9,14 +19,14 @@ from pydantic import BaseModel
 from twitter_langchain.twitter_api_wrapper import TwitterApiWrapper
 
 
-class TwitterAction(BaseTool):  # type: ignore[override]
+class TwitterTool(BaseTool):  # type: ignore[override]
     """Tool for interacting with the Twitter API."""
 
     twitter_api_wrapper: TwitterApiWrapper
-    mode: str
     name: str = ""
     description: str = ""
     args_schema: type[BaseModel] | None = None
+    func: Callable[..., str]
 
     def _run(
         self,
@@ -33,4 +43,4 @@ class TwitterAction(BaseTool):  # type: ignore[override]
             parsed_input_args = validated_input_data.model_dump()
         else:
             parsed_input_args = {"instructions": instructions}
-        return self.twitter_api_wrapper.run(self.mode, **parsed_input_args)
+        return self.twitter_api_wrapper.run_action(self.func, **parsed_input_args)